Rename module folder and clarify README outputs
- Rename human_in_the_loop_cdk/ to cdk/ to avoid name duplication
- Clarify deployment output as link to state machine in Step Functions console
