+++ human-in-the-loop-cdk/README.md
@@ -39,7 +39,7 @@ Important: this application uses various AWS services and there are costs associ
     ```bash
     cdk deploy --parameters ModeratorEmailAddress=your-email@example.com
     ```
-1. Note the outputs from the CDK deployment: the Step Functions console URL and the API Gateway endpoint URL.
+1. Note the outputs from the CDK deployment: the link to the state machine in the Step Functions console and the API Gateway endpoint URL.
 
 ## How it works
