+++ human-in-the-loop-cdk/app.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import aws_cdk as cdk
 
-from human_in_the_loop_cdk.human_in_the_loop_cdk_stack import HumanInTheLoopCdkStack
+from cdk.human_in_the_loop_cdk_stack import HumanInTheLoopCdkStack
 
 
 app = cdk.App()
